Option to run the dns proxy but not capture requests
This new option starts the DNS proxy as usual but does not install
the firewall rule to capture DNS queries.

The expectation is that a local DNS server, such as BIND, will direct
queries for the remote domain to the DNS proxy.  This can be done
in BIND with a forward zone declaration such as:

zone "foo.example.com" IN {
	type forward;
	forwarders port 12303 { 127.0.0.1; };
};

When sshuttle was started with:

$ sshuttle -l 12303 --dns-server-only -r user@host.example.com