Warn about --listen on all addresses (0.0.0.0) and --dns
When listening on all addresses (e.g. --listen 0.0.0.0) and using --dns,
DNS will stop working locally.

This happens because we install a REDIRECT rule that sends all outgoing
traffic to port 53 (--dport) to the sshuttle udp port on the loopback address.

Later when the response comes back from the server we use that same socket to
send the reply to the original application.

If the socket is bound to loopback, the NAT in iptables that matched earlier
will correctly translate back the source. However if the socket is bound
to all addresses, the response which has as destination the IP from where the
packet would have left the machine, won't be delivered to the application via
loopback but instead via the interface with that address.

Given the above, the iptables rule won't match and the source won't be
translated. The packet is delivered to the application with an unexpected
source and gets dropped.

Unfortunately this change will not solve this problem, but rather detect when
the user requests --dns and --listen on all interfaces, emit a warning and
make the DNS proxy bind to localhost. In this case DNS will continue to work
locally, however remote machines trying to redirect DNS traffic to sshuttle
won't be able to so anymore, unless --listen is provided with a specific IP
address.

I would be happy to implement any other ideas, I just can't think of anything
that would properly resolve this right now.

This is the issue reported in #157.