Restore partial subnet option parsing behavior
This restores behavior that was changed in c4a41ad which stopped `.0`
filling IP addresses correctly.

This change meant that entering in `10/8` as a subnet would end up being
parsed into `0.0.0.10/8` which is entirely different than `10.0.0.0/8`.

We are attempting to restore this behavior while still supporting
current behavior.

Fixes GH-184